Move flyway migrate function to plugin folder
